requirements.txt
Flask>=3.0.0
httpx>=0.27.0
python-dotenv>=1.0.0

# Development / testing
pytest>=8.0.0
